Добавляет контроллер TripController. Переносит в него логику рендера ивентов и дней. Перемещает функции из utils.js в common.js
